+++ skydrop/src/gui/gui_list.cpp
@@ -4,7 +4,7 @@
 gui_list_gen * gui_list_gen_f = NULL;
 gui_list_act * gui_list_act_f = NULL;
 
-uint16_t gui_list_size = 0;
+uint8_t gui_list_size = 0;
 uint8_t gui_list_back = 0;
 int16_t gui_list_y_offset = 0;
 uint8_t gui_list_index[NUMBER_OF_GUI_TASKS];
@@ -41,7 +41,7 @@ void gui_list_draw()
 	else
 		gui_list_middle_hold = task_get_ms_tick();
 
-	for (uint16_t i = 0; i < gui_list_size; i++)
+	for (uint8_t i = 0; i < gui_list_size; i++)
 	{
 		height = 1 + t_h;
